Move static funcs out to private free funcs
For the ISAN check I need to use the func in two cases, so the choice is
between private static or private free. I'm really not sure if there are
clear benefits of one over the other, but at least with the free func, I
can ommit the reference to Self, and perhaps that looks a little
cleaner.