Arithmetic safety
Quite embarrassingly, I accidentally was multiplying UInt8 numbers
and then converting to Int, rather than converting to Int and then
multiplying. This revealed itself as an arithmetic overflow in the
parsing of a SCTE-35 splice info section which has now been added
to the known test-cases.

This change makes sure that all locations where we do multiplication
is now converted to Int first. I also made a check for addition to
be safe.